==> proteomics/0-create_study_design_tables.Rmd <==
---
title: "Chr8 Study design tables"
output: html_document
---

```{r setup}
library(PNNL.DMS.utils)
library(stringr)
library(dplyr)
```



```{r}
path_to_global_study_design <- "study_design/global_study_design/"
path_to_global_with_mouse_study_design <- "study_design/global_with_mouse_study_design/"
path_to_phospho_study_design <- "study_design/phospho_study_design/"

dir.create(path_to_global_study_design)
dir.create(path_to_global_with_mouse_study_design)
dir.create(path_to_phospho_study_design)

global_data_package_num <- 5243
global_with_mouse_data_package_num <- 5343
phospho_data_package_num <- 5344

```


Samples, references and fractions table must be created from study-specific metadata. 

See `r help(read_study_design)` for more information.


```{r samples}
library(readxl)

metadataTable <- read_excel("data/Chr8-MetaDataSheet.xlsx")

# add Plex and Tube numbers
metadataTable$Plex <- 1
metadataTable$Tube <- NA
tube.num <- 0
for (i in 1:nrow(metadataTable)) {
  if (metadataTable$SampleID[i] != "reference") {
    tube.num <- tube.num + 1
    metadataTable$Tube[i] <- tube.num
  }
}

samples <- metadataTable %>% as.data.frame() %>%
  dplyr::select(Tube, ReporterName = Channel, 
                PlexID = Plex, 'SampleID') %>%
  mutate(QuantBlock = "1",
         ReporterAlias = as.character(Tube),
         PlexID = as.character(PlexID),
         ReporterAlias = case_when(is.na(Tube) ~ "ref",
                                   TRUE ~ ReporterAlias), # customize if reference changes
         ReporterName = case_when(ReporterName == "126C" ~ "126",       ## Here we make the channel names match the
                                  ReporterName == "134" ~ "134N",       ## format of the converter_ion conversion table.
                                  TRUE ~ ReporterName),
         MeasurementName = as.character(Tube)) # measurement names must be unique

samples_p <- samples
samples_p[grepl("Pooled", samples_p$'SampleID'), "MeasurementName"] <- c("Pooled 1", "Pooled 2")
samples_p[grepl("Pooled", samples_p$'SampleID'), "ReporterAlias"] <- c("Pooled 1", "Pooled 2")
samples_p[samples_p$ReporterAlias == "ref" | is.na(samples_p$ReporterAlias), "MeasurementName"] <- NA

samples[grepl("Pooled", samples$MeasurementName), "MeasurementName"] <- NA
samples[samples$ReporterAlias == "ref" | is.na(samples$ReporterAlias), "MeasurementName"] <- NA

write.table(samples %>% select(-Tube, -'SampleID'), 
            paste0(path_to_global_study_design, "samples.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples_p %>% select(-Tube, -'SampleID'), 
            paste0(path_to_global_study_design, "samples_pooled.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples %>% select(-Tube, -'SampleID'), 
            paste0(path_to_global_with_mouse_study_design, "samples.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples_p %>% select(-Tube, -'SampleID'), 
            paste0(path_to_global_with_mouse_study_design, "samples_pooled.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples %>% select(-Tube, -'SampleID'), 
            paste0(path_to_phospho_study_design, "samples.txt"),
            quote=F, sep="\t", row.names=F)

write.table(samples_p %>% select(-Tube, -'SampleID'), 
            paste0(path_to_phospho_study_design, "samples_pooled.txt"),
            quote=F, sep="\t", row.names=F)

```


```{r references}
references <- samples %>%
  dplyr::select(PlexID, QuantBlock) %>%
  unique() %>% mutate(Reference = "ref")

write.table(references, paste0(path_to_global_study_design, "references.txt"),
            quote=F, sep="\t", row.names=F)

write.table(references, paste0(path_to_global_with_mouse_study_design, "references.txt"),
            quote=F, sep="\t", row.names=F)

write.table(references, paste0(path_to_phospho_study_design, "references.txt"),
            quote=F, sep="\t", row.names=F)

```


```{r fractions}
## Global
job_records <- PNNL.DMS.utils::get_job_records_by_dataset_package(global_data_package_num)

# fractions <- data.frame(Dataset = unique(job_records$Dataset)) %>%
#   mutate(PlexID = sub("^PTRC_Exp22_Plex_0(.*)_G_.*$", "\\1", Dataset)) # customize if Dataset format changes

# only 1 plex so:
fractions <- data.frame(Dataset = unique(job_records$Dataset), PlexID = 1)

write.table(fractions, paste0(path_to_global_study_design, "fractions.txt"),
            quote=F, sep="\t", row.names=F)

## Global with mouse
job_records <- PNNL.DMS.utils::get_job_records_by_dataset_package(global_with_mouse_data_package_num)

# fractions <- data.frame(Dataset = unique(job_records$Dataset)) %>%
#   mutate(PlexID = sub("^PTRC_Exp22_Plex_0(.*)_G_.*$", "\\1", Dataset)) # customize if Dataset format changes

# only 1 plex so:
fractions <- data.frame(Dataset = unique(job_records$Dataset), PlexID = 1)

write.table(fractions, paste0(path_to_global_with_mouse_study_design, "fractions.txt"),
            quote=F, sep="\t", row.names=F)

## Phospho
job_records <- PNNL.DMS.utils::get_job_records_by_dataset_package(phospho_data_package_num)

# fractions <- data.frame(Dataset = unique(job_records$Dataset)) %>%
#   mutate(PlexID = sub("^PTRC_Exp22_Plex_0(.*)_P_.*$", "\\1", Dataset)) # customize if Dataset format changes

# only 1 plex so:
fractions <- data.frame(Dataset = unique(job_records$Dataset), PlexID = 1)

write.table(fractions, paste0(path_to_phospho_study_design, "fractions.txt"),
            quote=F, sep="\t", row.names=F)

```
