==> proteomics/3-normalize_and_batch_correction.Rmd <==
---
title: "Chr8 Phospho proteomics"
output: html_document
---

Preliminary summary




```{r include=FALSE}
library(MSnSet.utils)
library(dplyr)
library(ggplot2)
library(readxl)

global_crosstab <- read.table("../proteomics/data/global_data/Chr8_crosstab_global_gene_original.txt")
global_with_mouse_crosstab <- read.table("../proteomics/data/global_with_mouse_data/Chr8_crosstab_global_old_database_gene_original.txt")
phospho_crosstab <- read.table("../proteomics/data/phospho_data/Chr8_crosstab_phospho_siteID_original.txt")
phospho_crosstab_prot <- read.table("../proteomics/data/phospho_data/Chr8_crosstab_phospho_siteID_prot_original.txt", sep = "\t")
phospho_pep_crosstab <- read.table("../proteomics/data/phospho_data/Chr8_crosstab_phospho_peptide_original.txt")

meta <- read_excel("../proteomics/data/Chr8-MetaDataSheet.xlsx") 

# add Plex and Tube numbers
meta$Plex <- 1
meta$Tube <- NA
tube.num <- 0
for (i in 1:nrow(meta)) {
  if (meta$SampleID[i] != "reference") {
    tube.num <- tube.num + 1
    meta$Tube[i] <- tube.num
  }
}

meta <- meta %>%
  filter(!is.na(Tube)) %>%
  mutate(Tube = as.character(Tube),
         Plex = as.character(Plex)) %>% as.data.frame()
#rownames(meta) <- meta$Tube # duplicate 'row.names' are not allowed (because of EMPTY and PNNL REFERENCE rows)

# remove EMPTY rows
meta <- meta[meta$Tube != "EMPTY", ]

# fix sample IDs
meta[meta$SampleID == "WU-22t Cshim81", ]$SampleID <- "WU-225"
meta[meta$SampleID == "WU-487CPP1", ]$SampleID <- "WU-487"

# add Chr8 status
meta$Chr8_amplified <- TRUE
non.chr8.amp <- c("JH-2-055", "WU-487")
meta[meta$SampleID %in% non.chr8.amp, ]$Chr8_amplified <- FALSE
meta$Chr8_strongly_amplified <- TRUE
meta[meta$SampleID %in% non.chr8.amp, ]$Chr8_strongly_amplified <- FALSE
meta[meta$SampleID == "JH-2-002", ]$Chr8_strongly_amplified <- NA

colnames(global_crosstab) <- sub("^X", "", colnames(global_crosstab))
colnames(global_with_mouse_crosstab) <- sub("^X", "", colnames(global_with_mouse_crosstab))
colnames(phospho_crosstab) <- sub("^X", "", colnames(phospho_crosstab))
colnames(phospho_pep_crosstab) <- sub("^X", "", colnames(phospho_pep_crosstab))
colnames(phospho_crosstab_prot) <- sub("^data.", "", colnames(phospho_crosstab_prot))

## data only
phospho_prot_mat <- phospho_crosstab_prot[, intersect(as.character(1:ncol(phospho_crosstab_prot)), colnames(phospho_crosstab_prot))]


```


Missing data


```{r}
global_missing <- rowSums(is.na(global_crosstab))
global_with_mouse_missing <- rowSums(is.na(global_with_mouse_crosstab))
phospho_missing <- rowSums(is.na(phospho_crosstab))
phospho_pep_missing <- rowSums(is.na(phospho_pep_crosstab))
phospho_prot_missing <- rowSums(is.na(phospho_prot_mat))

hist(global_missing)
hist(global_with_mouse_missing)
hist(phospho_missing)
hist(phospho_pep_missing)
hist(phospho_prot_missing)

## 50 % missingness filter, lose 178 global features and 27205 phosphosites
global_crosstab <- global_crosstab[global_missing < ncol(global_crosstab)/2, ]
global_with_mouse_crosstab <- global_with_mouse_crosstab[global_with_mouse_missing < ncol(global_with_mouse_crosstab)/2, ]
phospho_crosstab <- phospho_crosstab[phospho_missing < ncol(phospho_crosstab)/2, ]
phospho_pep_crosstab <- phospho_pep_crosstab[phospho_pep_missing < ncol(phospho_pep_crosstab)/2, ]
phospho_prot_mat <- phospho_prot_mat[phospho_prot_missing < ncol(phospho_crosstab)/2, ]

```


Raw data stats.


```{r}
## row medians
hist(apply(global_crosstab, 1, median, na.rm = T))
hist(apply(global_with_mouse_crosstab, 1, median, na.rm = T))
hist(apply(phospho_crosstab, 1, median, na.rm = T))
hist(apply(phospho_pep_crosstab, 1, median, na.rm = T))
hist(apply(phospho_prot_mat, 1, median, na.rm = T))

global_row_medians <- apply(global_crosstab, 1, median, na.rm = T)
global_with_mouse_row_medians <- apply(global_with_mouse_crosstab, 1, median, na.rm = T)
phospho_row_medians <- apply(phospho_crosstab, 1, median, na.rm = T)
phospho_pep_row_medians <- apply(phospho_pep_crosstab, 1, median, na.rm = T)
phospho_prot_row_medians <- apply(phospho_prot_mat, 1, median, na.rm = T)

global_crosstab <- sweep(global_crosstab, 1, global_row_medians, FUN = '-')
global_with_mouse_crosstab <- sweep(global_with_mouse_crosstab, 1, global_with_mouse_row_medians, FUN = '-')
phospho_crosstab <- sweep(phospho_crosstab, 1, phospho_row_medians, FUN = '-')
phospho_pep_crosstab <- sweep(phospho_pep_crosstab, 1, phospho_pep_row_medians, FUN = '-')
phospho_prot_mat <- sweep(phospho_prot_mat, 1, phospho_prot_row_medians, FUN = '-')

```



```{r}
table(colnames(global_crosstab) == colnames(phospho_crosstab))
table(colnames(global_with_mouse_crosstab) == colnames(phospho_crosstab))
table(colnames(global_crosstab) == colnames(phospho_pep_crosstab))
table(colnames(global_crosstab) == colnames(phospho_prot_mat))

## sample medians
hist(apply(global_crosstab, 2, median, na.rm = T))
hist(apply(global_with_mouse_crosstab, 2, median, na.rm = T))
hist(apply(phospho_crosstab, 2, median, na.rm = T))
hist(apply(phospho_pep_crosstab, 2, median, na.rm = T))
hist(apply(phospho_prot_mat, 2, median, na.rm = T))

## Using global coefficients to normalize phospho.
global_sample_coef <- apply(global_crosstab, 2, median, na.rm = T)
global_with_mouse_sample_coef <- apply(global_with_mouse_crosstab, 2, median, na.rm = T)

global_crosstab <- sweep(global_crosstab, 2, global_sample_coef, FUN = '-')
global_with_mouse_crosstab <- sweep(global_with_mouse_crosstab, 2, global_with_mouse_sample_coef, FUN = '-')
phospho_crosstab <- sweep(phospho_crosstab, 2, global_sample_coef, FUN = '-')
phospho_pep_crosstab <- sweep(phospho_pep_crosstab, 2, global_sample_coef, FUN = '-')
phospho_prot_mat <- sweep(phospho_prot_mat, 2, global_sample_coef, FUN = '-')

```



```{r}
m_global <- MSnSet(exprs = global_crosstab %>% as.matrix(), 
                   pData = meta[colnames(global_crosstab), ])

m_global_with_mouse <- MSnSet(exprs = global_with_mouse_crosstab %>% as.matrix(), 
                   pData = meta[colnames(global_with_mouse_crosstab), ])

m_phospho <- MSnSet(exprs = phospho_crosstab %>% as.matrix(), 
                    pData = meta[colnames(phospho_crosstab), ])

m_phospho_pep <- MSnSet(exprs = phospho_pep_crosstab %>% as.matrix(), 
                    pData = meta[colnames(phospho_pep_crosstab), ])

m_phospho_prot <- MSnSet(exprs = phospho_prot_mat %>% as.matrix(), 
                         pData = meta[colnames(phospho_prot_mat), ])

```


Both datasets show strong batch effect


```{r}
phenos <- c("Plex", "SampleID", "Chr8_amplified", "Chr8_strongly_amplified")
for (i in 1:length(phenos)) {
  plot_pca(m_global, phenotype = phenos[i]) + ggtitle("Global PCA")
  ggsave(paste0("Chr8_global_PCA_by_", phenos[i], ".pdf"))
  
  plot_pca(m_global_with_mouse, phenotype = phenos[i]) + ggtitle("Global PCA")
  ggsave(paste0("Chr8_global_old_database_PCA_by_", phenos[i], ".pdf"))
  
  plot_pca(m_phospho, phenotype = phenos[i]) + ggtitle("Phospho PCA")
  ggsave(paste0("Chr8_phospho_PCA_by_", phenos[i], ".pdf"))
  
  plot_pca(m_phospho_pep, phenotype = phenos[i]) + ggtitle("Phospho PCA")
  ggsave(paste0("Chr8_phospho_peptide_PCA_by_", phenos[i], ".pdf"))
  
  plot_pca(m_phospho_prot, phenotype = phenos[i]) + ggtitle("Phospho prot PCA")
  ggsave(paste0("Chr8_phospho_prot_PCA_by_", phenos[i], ".pdf"))
}

```


```{r}
# can only perform correction for plex if more than one plex
if (length(unique(meta$Plex)) > 1) {
  m_global_corrected <- correct_batch_effect_NA(m_global, "Plex", par.prior = T)
  m_global_with_mouse_corrected <- correct_batch_effect_NA(m_global_with_mouse, "Plex", par.prior = T)
  m_phospho_corrected <- correct_batch_effect_NA(m_phospho, "Plex", par.prior = T)
  m_phospho_pep_corrected <- correct_batch_effect_NA(m_phospho_pep, "Plex", par.prior = T)
  
  m_phospho_prot_corrected <- correct_batch_effect_NA(m_phospho_prot, "Plex", par.prior = T)
} else {
  m_global_corrected <- m_global
  m_global_with_mouse_corrected <- m_global_with_mouse
  m_phospho_corrected <- m_phospho
  m_phospho_pep_corrected <- m_phospho_pep
  
  m_phospho_prot_corrected <- m_phospho_prot
}
```



```{r}
phenos <- c("Plex", "SampleID", "Chr8_amplified", "Chr8_strongly_amplified")
for (i in 1:length(phenos)) {
  plot_pca(m_global, phenotype = phenos[i]) + ggtitle("Global PCA")
  ggsave(paste0("Chr8_global_PCA_by_", phenos[i], "_corrected.pdf"))
  
  plot_pca(m_global_with_mouse, phenotype = phenos[i]) + ggtitle("Global PCA")
  ggsave(paste0("Chr8_global_old_database_PCA_by_", phenos[i], "_corrected.pdf"))
  
  plot_pca(m_phospho, phenotype = phenos[i]) + ggtitle("Phospho PCA")
  ggsave(paste0("Chr8_phospho_PCA_by_", phenos[i], "_corrected.pdf"))
  
  plot_pca(m_phospho_pep, phenotype = phenos[i]) + ggtitle("Phospho PCA")
  ggsave(paste0("Chr8_phospho_peptide_PCA_by_", phenos[i], "_corrected.pdf"))
  
  plot_pca(m_phospho_prot, phenotype = phenos[i]) + ggtitle("Phospho prot PCA")
  ggsave(paste0("Chr8_phospho_prot_PCA_by_", phenos[i], "_corrected.pdf"))
}

```

Push to synapse.

```{r}
write.table(exprs(m_global_corrected), 
            file = "data/global_data/Chr8_crosstab_global_gene_corrected.txt",
            quote=F, sep="\t")

write.table(exprs(m_global_with_mouse_corrected), 
            file = "data/global_with_mouse_data/Chr8_crosstab_global_old_database_gene_corrected.txt",
            quote=F, sep="\t")

write.table(exprs(m_phospho_corrected), 
            file = "data/phospho_data/Chr8_crosstab_phospho_SiteID_corrected.txt",
            quote=F, sep="\t")

write.table(exprs(m_phospho_pep_corrected), 
            file = "data/phospho_data/Chr8_crosstab_phospho_peptide_corrected.txt",
            quote=F, sep="\t")


phospho_prot_corrected <- exprs(m_phospho_prot_corrected) %>%
  as.data.frame() %>%
  mutate(feature = rownames(.))
mapping_df <- phospho_crosstab_prot %>%
  mutate(feature = rownames(.)) %>%
  select(feature, Peptide_kstar, Protein_kstar, SiteID)

phospho_prot_corrected <- merge(phospho_prot_corrected, mapping_df, by = "feature")
rownames(phospho_prot_corrected) <- phospho_prot_corrected$feature
phospho_prot_corrected <- phospho_prot_corrected %>% select(-feature) %>%
  select(SiteID, Peptide_kstar, Protein_kstar, everything())

colnames(phospho_prot_corrected) <- sub("(^[0-9].*$)", "data:\\1", colnames(phospho_prot_corrected))
write.table(phospho_prot_corrected, 
            file = "data/phospho_data/Chr8_crosstab_phospho_SiteID_prot_corrected.txt",
            sep="\t")

```


```{r}
# can only perform if more than one plex
if (length(unique(meta$Plex)) > 1) {
  plex_global <- limma_gen(m_global, "~Plex", "Plex")
  plex_global_with_mouse <- limma_gen(m_global_with_mouse, "~Plex", "Plex")
  plex_phospho <- limma_gen(m_phospho, "~Plex", "Plex")
  plex_phospho_pep <- limma_gen(m_phospho_pep, "~Plex", "Plex")
  
  plex_global_corrected <- limma_gen(m_global_corrected, "~Plex", "Plex")
  plex_global_with_mouse_corrected <- limma_gen(m_global_with_mouse_corrected, "~Plex", "Plex")
  plex_phospho_corrected <- limma_gen(m_phospho_corrected, "~Plex", "Plex")
  plex_phospho_pep_corrected <- limma_gen(m_phospho_pep_corrected, "~Plex", "Plex")
}

```



```{r}
# can only perform if more than one plex
if (length(unique(meta$Plex)) > 1) {
  hist(plex_global$P.Value)
  hist(plex_global_corrected$P.Value)
  
  hist(plex_global_with_mouse$P.Value)
  hist(plex_global_with_mouse_corrected$P.Value)
  
  hist(plex_phospho$P.Value)
  hist(plex_phospho_corrected$P.Value)
  
  hist(plex_phospho_pep$P.Value)
  hist(plex_phospho_pep_corrected$P.Value)
}
```





```{r}
library(readxl)
## Finalizing KSTAR input
crosstab <- read.table(file = "data/phospho_data/Chr8_crosstab_phospho_SiteID_prot_corrected.txt",
            sep="\t")
colnames(crosstab) <- sub("data.", "", colnames(crosstab))
meta <- read_excel("../proteomics/data/Chr8-MetaDataSheet.xlsx") 

# add Plex and Tube numbers
meta$Plex <- 1
meta$Tube <- NA
tube.num <- 0
for (i in 1:nrow(meta)) {
  if (meta$SampleID[i] != "reference") {
    tube.num <- tube.num + 1
    meta$Tube[i] <- tube.num
  }
}

meta <- meta %>%
  filter(!is.na(Tube)) %>%
  mutate(Tube = as.character(Tube),
         Plex = as.character(Plex)) %>% as.data.frame()
#rownames(meta) <- meta$Tube

# remove EMPTY rows
#meta <- meta[meta$Tube != "EMPTY", ]

# add Treatment column
meta$Treatment <- "Chr8"
#meta[meta$SampleID == "reference", ]$Treatment <- "reference"

crosstab_kstar <- crosstab %>%
  select(SiteID, Peptide_kstar, Protein_kstar)

# no "baseline" aka control for Chr8 samples
all_groups <- na.omit(unique(meta$Treatment))
for (Treatment in all_groups){
  samples <- meta %>% filter(Treatment == Treatment) %>%
    pull(Tube)
  mat <- crosstab[, na.omit(samples)]
  crosstab_kstar[paste0("data:", Treatment, "_logFC")] <- apply(mat, 1, mean, na.rm = T)
}

data_folder = "data/phospho_data/"
crosstab_prefix = "Chr8"
## showing logFC vs baseline and none for each treatment.
write.table(crosstab_kstar,
            file=file.path(data_folder, paste(crosstab_prefix, "crosstab_phospho_treatments_logFC.txt", sep="_")),
            sep="\t", quote = F)

```
